+++ Player/Scan/ScanNodeCreator.cs
@@ -1,5 +1,4 @@
 ﻿using LethalMDK;
-using LethalMDK.World;
 using UnityEngine;
 using UnityMDK.Injection;
 
@@ -21,20 +20,20 @@ public class ScanNodeCreator : MonoBehaviour
         bool isScrap = item.itemProperties.isScrap;
         
         GameObject scanNode = new("Scan Node") { layer = Layers.ScanNode };
+        
+        scanNode.transform.SetParent(transform);
+        scanNode.transform.SetLocalPositionAndRotation(Vector3.zero, Quaternion.identity);
 
-        DisableableScanNodeProperties properties = scanNode.AddComponent<DisableableScanNodeProperties>();
+        CustomScanNodeProperties properties = scanNode.AddComponent<CustomScanNodeProperties>();
         properties.headerText = item.itemProperties.itemName;
         properties.subText = isScrap ? $"Value: ${item.scrapValue}" : "???";
         properties.minRange = 1;
         properties.maxRange = 13;
         properties.creatureScanID = -1;
-        properties.nodeType = (int)(isScrap ? ScanNodes.EType.Scrap : ScanNodes.EType.Utility);
+        properties.nodeType = isScrap ? PingScan.NodeScrap : PingScan.NodeUtility;
         properties.requiresLineOfSight = true;
 
         BoxCollider collider = scanNode.AddComponent<BoxCollider>();
         collider.size = Vector3.one * 0.2f;
-        
-        scanNode.transform.SetParent(transform);
-        scanNode.transform.SetLocalPositionAndRotation(Vector3.zero, Quaternion.identity);
     }
 }